new,title: introduce new title format
After real life practice, I found that the single letter prefix is
nonsensical for guest. It is unnecessary effort to understand my commit.
So I decide to use three letter to describe the commit type.

Besides, a capital letter is really ugly. And path divider blend the
prefix and module path, which makes the commit title hard to read. So
here I use `,` to replace it.

Signed-off-by: Avimitin <avimitin@gmail.com>