rew: update prefix keyword and divider
= Major changes
The prefix keywords `rft` and `chr` confuse guest and me. So I decide to
replace it will `rew` which stands for `rewrite` and odd.

The reason about I use rewrite not refactor is described in the main
document.

Also I've update the reference format. ASCIIDOC format is too verbose in
commit message.

Signed-off-by: Avimitin <avimitin@gmail.com>